+++ src/components/organisms/NavBarSchedule/NavBarSchedule.scss
@@ -6,6 +6,8 @@ $NavBarSchedule--padding: 20px;
 $weekday--padding: 8px 10px;
 $weekday--margin-right: 10px;
 
+$schedule-btn--padding: 0.35em 1em;
+
 .NavBarSchedule {
   @include scrollbar;
   position: fixed;
@@ -35,18 +37,28 @@ $weekday--margin-right: 10px;
   }
 
   &__weekdays {
+    @include scrollbar;
     display: flex;
-    padding-left: $NavBarSchedule--padding;
+    overflow: auto;
+    cursor: default;
   }
 
   &__download-buttons {
     display: flex;
+    position: fixed;
+    left: 0;
+    right: 0;
+    bottom: 0;
+    justify-content: flex-end;
+    align-items: center;
+    background-color: opaque-black(0.9);
+    height: 40px;
     padding-left: $NavBarSchedule--padding;
 
     margin: $spacing--md 0;
   }
 
-  &__download-schedule-btn {
+  & .NavBarSchedule__download-schedule-btn {
     cursor: pointer;
     opacity: 0.8;
 
@@ -54,6 +66,7 @@ $weekday--margin-right: 10px;
 
     text-shadow: 0 2px 4px opaque-black(0.3);
     box-shadow: 0 3px 6px opaque-black(0.2);
+    padding: $schedule-btn--padding;
 
     &:hover,
     &:focus {
@@ -66,7 +79,6 @@ $weekday--margin-right: 10px;
     cursor: pointer;
     padding: $weekday--padding;
     margin-right: $weekday--margin-right;
-    border-radius: $border-radius--xl;
     opacity: 0.8;
 
     &-button {
@@ -77,15 +89,25 @@ $weekday--margin-right: 10px;
       &:focus {
         color: $white;
         opacity: 1;
+        outline: none;
       }
     }
 
     &--active {
       opacity: 1;
-      background-color: $primary;
+      border-bottom: 1px solid white;
       text-shadow: 0 2px 4px opaque-black(0.3);
       box-shadow: 0 3px 6px opaque-black(0.2);
-      border: none;
+    }
+
+    &--disabled {
+      pointer-events: none;
+      cursor: default;
+
+      button {
+        cursor: default;
+        color: rgba(255, 255, 255, 0.383);
+      }
     }
 
     &:hover {
