--- src/components/organisms/NavBarScheduleV2/NavBarSchedule.scss
@@ -1,108 +1,0 @@
-@import "scss/constants.scss";
-
-$NavBarScheduleV2--height: 620px;
-$NavBarScheduleV2--padding: 20px;
-
-$weekday--padding: 8px 10px;
-$weekday--margin-right: 10px;
-
-.NavBarScheduleV2 {
-  @include scrollbar;
-  position: fixed;
-  z-index: z(navbar-schedule);
-  top: 0;
-  left: 0;
-  width: 100%;
-  height: $NavBarScheduleV2--height;
-
-  padding-top: $navbar-height;
-  padding-left: $NavBarScheduleV2--padding;
-
-  background-color: opaque-black(0.9);
-  overflow: auto;
-
-  pointer-events: none;
-  opacity: 0;
-  transform: translateY(-$NavBarScheduleV2--height);
-  backdrop-filter: blur(5px);
-  box-shadow: 0 20px 50px 0 rgba(0, 0, 0, 0.33);
-  transition: all 400ms $transition-function;
-
-  &--show {
-    pointer-events: all;
-    opacity: 1;
-    transform: translateY(0);
-  }
-
-  &__weekdays {
-    display: flex;
-    padding-left: $NavBarScheduleV2--padding;
-    width: 500px;
-    overflow: scroll;
-    cursor: default;
-  }
-
-  &__download-buttons {
-    display: flex;
-    padding-left: $NavBarScheduleV2--padding;
-
-    margin: $spacing--md 0;
-  }
-
-  &__download-schedule-btn {
-    cursor: pointer;
-    opacity: 0.8;
-
-    margin-right: $spacing--sm;
-
-    text-shadow: 0 2px 4px opaque-black(0.3);
-    box-shadow: 0 3px 6px opaque-black(0.2);
-
-    &:hover,
-    &:focus {
-      opacity: 1;
-    }
-  }
-
-  &__weekday {
-    display: inline;
-    cursor: pointer;
-    padding: $weekday--padding;
-    margin-right: $weekday--margin-right;
-    opacity: 0.8;
-
-    &-button {
-      border: none;
-      color: $white;
-      background: inherit;
-      outline: none;
-
-      &:focus {
-        color: $white;
-        opacity: 1;
-        outline: none;
-      }
-    }
-
-    &--active {
-      opacity: 1;
-      border-bottom: 1px solid white;
-      text-shadow: 0 2px 4px opaque-black(0.3);
-      box-shadow: 0 3px 6px opaque-black(0.2);
-    }
-
-    &--disabled {
-      pointer-events: none;
-      cursor: default;
-
-      button {
-        cursor: default;
-        color: rgba(255, 255, 255, 0.383);
-      }
-    }
-
-    &:hover {
-      opacity: 1;
-    }
-  }
-}
